confluencebot/pyproject.toml
[tool.poetry]
name = "confluencebot"
version = "0.1.0"
description = ""
authors = ["Justin Levi Winter <justinlevi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
streamlit = "^1.22.0"
watchdog = "^3.0.0"
pypdf2 = "^3.0.1"
openai = "^0.27.6"
tiktoken = "^0.4.0"
langchain = "0.0.189"
chromadb = "0.3.25"
pytesseract = "0.3.10"
atlassian-python-api = "3.38.0"
beautifulsoup4 = "^4.12.2"
debugpy = "^1.6.7"
lxml = "^4.6.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
mock = "^5.0.2"

[dev-dependencies]
streamlit = "*"
PyPDF2 = "*"
langchain = "*"
tiktoken = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
